+++ app/lib/static-lookups.lib.js
@@ -271,6 +271,15 @@ const naldRegions = {
   YO: 'Yorkshire'
 }
 
+const NoticeJourney = Object.freeze({ STANDARD: 'standard', ALERTS: 'alerts' })
+
+const NoticeType = Object.freeze({
+  ABSTRACTION_ALERTS: 'abstractionAlerts',
+  INVITATIONS: 'invitations',
+  PAPER_RETURN: 'returnForms',
+  REMINDERS: 'reminders'
+})
+
 const organisationTypes = ['individual', 'limitedCompany', 'limitedLiabilityPartnership', 'publicLimitedCompany']
 
 /**
@@ -453,6 +462,8 @@ module.exports = {
   flowUnits,
   naldAreaCodes,
   naldRegions,
+  NoticeJourney,
+  NoticeType,
   organisationTypes,
   returnCycleDates,
   returnPeriodDates,
